Chain area based stimate delete parent (#1965)
* delete nodeDef with areaBased

* Update webapp/components/survey/NodeDefDetails/AnalysisProps/AreaBasedEstimated/AreaBasedEstimated.js
