+++ webapp/components/survey/NodeDefDetails/AnalysisProps/AreaBasedEstimated/index.js
@@ -1,0 +1,1 @@
+export { default } from './AreaBasedEstimated'
