+++ webapp/components/survey/NodeDefDetails/AnalysisProps/RScriptEditor/index.js
@@ -1,0 +1,1 @@
+export { default } from './RScriptEditor'
